% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fns.R
\name{doBRR}
\alias{doBRR}
\title{Function for doing brr}
\usage{
doBRR(
  df,
  fit,
  nbrr,
  us = 1:30,
  uPSU = 1:2,
  strata = NULL,
  PSU = NULL,
  weights = NULL
)
}
\arguments{
\item{df}{data frame containing all variables used in initial fitting, plus survey variables (strata, PSU, weights)}

\item{fit}{Initial fit on the full data, pass the formula into [mgcv::gam()]}

\item{nbrr}{Number of BRR samples}

\item{us}{Unique strata identifiers (corresponds to the unique levels of the variable specified by the strata argument)}

\item{uPSU}{Unique PSU identifiers (corresponds to the unique levels of the variable specified by the PSU argument)}

\item{strata}{Character vector of length 1 indicating the variable name for the strata identifier in the data frame used for model fitting}

\item{PSU}{Character vector of length 1 indicating the variable name for the PSU identifier in the data frame used for model fitting}

\item{weights}{Character vector of length 1 indicating the variable name for individuals' survey weights in the data frame used for model fitting}
}
\description{
Function for doing brr
}
